Revert "Fix version overflow"

+++ Telegram/SourceFiles/core/update_checker.cpp
@@ -402,7 +402,7 @@ bool UnpackUpdate(const QString &filepath) {
 
 	tempDir.mkdir(tempDir.absolutePath());
 
-	quint64 version;
+	quint32 version;
 	{
 		QDataStream stream(uncompressed);
 		stream.setVersion(QDataStream::Qt_5_1);
@@ -424,7 +424,7 @@ bool UnpackUpdate(const QString &filepath) {
 				LOG(("Update Error: downloaded alpha version %1 is not greater, than mine %2").arg(alphaVersion).arg(cAlphaVersion()));
 				return false;
 			}
-		} else if (int64(version) <= AppVersion) {
+		} else if (int32(version) <= AppVersion) {
 			LOG(("Update Error: downloaded version %1 is not greater, than mine %2").arg(version).arg(AppVersion));
 			return false;
 		}
